IMPORTANT RESULT: Tried working on MCStep for hours and have struggled with vectorising a random access of a grid where each step depends on the previous random selection. Its basically impossible to do without either removing randomness or removing accuracy as of right now so im going to move on for now.
I created a vectorisied method of finding the order of the program which also removes some calculations that were unecessary i.e. any calculation with the layer of zeroes in the looping. My way I believe seems very non mathematical but as can be seen by this version of the code results in the same eigenvalues.

I have now realised I should produce image evidence of these results so will do after this for this result and the vectorisation of the energies. My next steps are to look for other things than MCStep to vectorise then move on to another speed up method.